chore: rerun command refactor and improve dotnet test invocation
