+++ test/dotnet-test-rerun.UnitTests/Exceptions/RerunExceptionTests.cs
@@ -1,4 +1,5 @@
 ﻿using dotnet.test.rerun;
+using dotnet.test.rerun.RerunCommand;
 using FluentAssertions;
 using Xunit;
 
@@ -10,7 +11,7 @@ public class RerunExceptionTests
     public void RerunExceptionTests_ConstructorWithoutMessage_ThrowsException()
     {
         //Arrange
-        var message = "Exception of type 'dotnet.test.rerun.RerunException' was thrown.";
+        var message = "Exception of type 'dotnet.test.rerun.RerunCommand.RerunException' was thrown.";
         
         //Act
         var exception = new RerunException();
